+++ PlaygroundDotNetAPI.Tests/PokedexControllerTest.cs
@@ -1,0 +1,32 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
+using Moq;
+using PlaygroundDotNetAPI.Controllers;
+
+namespace PlaygroundDotNetAPI.Tests
+{
+    [TestFixture]
+    public class PokedexControllerTests
+    {
+        private PokedexController _controller;
+
+        [SetUp]
+        public void Setup()
+        {
+            //_controller = new PokedexController();
+        }
+
+        [Test]
+        public void TODO()
+        {
+            Assert.Fail("Check headers for EVERY call");
+            Assert.Fail("Count the items");
+            Assert.Fail("Should be 200");
+            Assert.Fail("Should be of type Pokemon");
+        }
+    }
+}
