+++ PlaygroundDotNetAPI.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34728.123
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PlaygroundDotNetAPI", "PlaygroundDotNetAPI\PlaygroundDotNetAPI.csproj", "{6CF740DB-20F3-4473-9B5C-3C3C000094B1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PlaygroundDotNetAPI.Tests", "PlaygroundDotNetAPI.Tests\PlaygroundDotNetAPI.Tests.csproj", "{E5FF7A73-454A-4858-A89A-5FB611F4A0BA}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6CF740DB-20F3-4473-9B5C-3C3C000094B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6CF740DB-20F3-4473-9B5C-3C3C000094B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6CF740DB-20F3-4473-9B5C-3C3C000094B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6CF740DB-20F3-4473-9B5C-3C3C000094B1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E5FF7A73-454A-4858-A89A-5FB611F4A0BA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E5FF7A73-454A-4858-A89A-5FB611F4A0BA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E5FF7A73-454A-4858-A89A-5FB611F4A0BA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E5FF7A73-454A-4858-A89A-5FB611F4A0BA}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {32337A10-9680-44FE-BDD7-891BC3A82FDC}
+	EndGlobalSection
+EndGlobal
