+++ PlaygroundDotNetAPI/PlaygroundDotNetAPI.http
@@ -1,0 +1,6 @@
+@PlaygroundDotNetAPI_HostAddress = http://localhost:5120
+
+GET {{PlaygroundDotNetAPI_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
