+++ PlaygroundDotNetAPI/appsettings.Development.json
@@ -1,0 +1,14 @@
+{
+    "Logging": {
+        "LogLevel": {
+            "Default": "Information",
+            "Microsoft.AspNetCore": "Warning"
+        }
+    },
+    "Db": {
+        "Type": "sqlite"
+    },
+    "ConnectionStrings": {
+        "DefaultConnection": "Data Source=Data/sqlite.db"
+    }
+}
