--- playground-dotnet-api.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 25.0.1706.3
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "playground-dotnet-api", "playground-dotnet-api\playground-dotnet-api.csproj", "{E6D5FD10-F3BA-4A17-9483-F9403F75EFE5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E6D5FD10-F3BA-4A17-9483-F9403F75EFE5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E6D5FD10-F3BA-4A17-9483-F9403F75EFE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E6D5FD10-F3BA-4A17-9483-F9403F75EFE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E6D5FD10-F3BA-4A17-9483-F9403F75EFE5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6A7634CD-14A8-4144-AFF3-1ECEFB118A42}
-	EndGlobalSection
-EndGlobal
