--- playground-dotnet-api/appsettings.Development.json
@@ -1,16 +1,0 @@
-﻿{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "Db": {
-    "Type": "sqlite"
-  },
-  "ConnectionStrings": {
-    "DefaultConnection": "Data Source=Data/sqlite.db"
-  },
-  "AllowedHosts": "*",
-  "AllowedOrigins": "*"
-}
