--- playground-dotnet-api/appsettings.json
@@ -1,11 +1,0 @@
-﻿{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "AllowedOrigins": "*"
-}
-
