--- playground-dotnet-api/web.config
@@ -1,13 +1,0 @@
-<?xml version="1.0"?>
-<configuration>
-    <system.webServer>
-		<security>
-			<requestFiltering removeServerHeader="true" />
-		</security>
-		<httpProtocol>
-			<customHeaders>
-				<remove name="X-Powered-By" />
-			</customHeaders>
-		</httpProtocol>
-    </system.webServer>
-</configuration>
